EY-2660 Fjernet ubrukte endepunkter for aareg og inntekskomponenten
- Disse brukes ikke, og kan også integreres med uten bruk av vår proxy
- La til konfigurasjon for tilbakekreving i nais-filene
