+++ apps/etterlatte-proxy/.nais/dev.yaml
@@ -41,24 +41,20 @@ spec:
   azure:
     application:
       enabled: true
-  tokenx:
-    enabled: true
   webproxy: true
   env:
     - name: STS_REST_URL
       value: http://security-token-service.default.svc.nais.local/rest/v1/sts/token
     - name: STS_SOAP_URL
       value: https://sts-q1.preprod.local/SecurityTokenServiceProvider/
-    - name: INNTEKTSKOMPONENTEN_URL
-      value: https://app-q2.adeo.no/inntektskomponenten-ws/rs/api/v1/hentinntektliste
-    - name: AAREG_URL
-      value: https://aareg-services.dev.intern.nav.no/api/v2
     - name: REGOPPSLAG_URL
       value: https://regoppslag.dev.intern.nav.no/rest
     - name: INSTITUSJONSOPPHOLD_URL
       value: https://inst2-q2.dev.intern.nav.no
     - name: INSTITUSJONSOPPHOLD_AZURE_SCOPE
       value: api://dev-fss.team-rocket.inst2-q2/.default
+    - name: TILBAKEKREVING_URL
+      value: https://cics-q1.adeo.no/oppdrag/K231CW80
   prometheus:
     enabled: true
     path: /internal/prometheus
@@ -74,12 +70,12 @@ spec:
         - application: sjekk-adressebeskyttelse
           namespace: etterlatte
           cluster: dev-gcp
-        - application: etterlatte-pdltjenester
-          namespace: etterlatte
-          cluster: dev-gcp
         - application: etterlatte-brev-api
           namespace: etterlatte
           cluster: dev-gcp
         - application: etterlatte-institusjonsopphold
+          namespace: etterlatte
+          cluster: dev-gcp
+        - application: etterlatte-tilbakekreving
           namespace: etterlatte
           cluster: dev-gcp
