+++ apps/etterlatte-proxy/src/main/kotlin/auth/Authentication.kt
@@ -10,14 +10,7 @@ import no.nav.etterlatte.config.Config
 import java.net.URL
 import java.util.concurrent.TimeUnit
 
-fun Application.installAuthentication(configAad: Config.AAD, configTokeX: Config.TokenX) {
-    val jwkProvider = JwkProviderBuilder(URL(configTokeX.metadata.jwksUri))
-        // cache up to 10 JWKs for 24 hours
-        .cached(10, 24, TimeUnit.HOURS)
-        // if not cached, only allow max 10 different keys per minute to be fetched from external provider
-        .rateLimited(10, 1, TimeUnit.MINUTES)
-        .build()
-
+fun Application.installAuthentication(configAad: Config.AAD) {
     val jwkProviderAad = JwkProviderBuilder(URL(configAad.metadata.jwksUri))
         // cache up to 10 JWKs for 24 hours
         .cached(10, 24, TimeUnit.HOURS)
@@ -26,22 +19,6 @@ fun Application.installAuthentication(configAad: Config.AAD, configTokeX: Config
         .build()
 
     install(Authentication) {
-        jwt("tokenX") {
-            verifier(jwkProvider, configTokeX.metadata.issuer)
-            validate { credentials ->
-                try {
-                    requireNotNull(credentials.payload.audience) {
-                        "Auth: Missing audience in token"
-                    }
-                    require(credentials.payload.audience.contains(configTokeX.clientId)) {
-                        "Auth: Valid audience not found in claims"
-                    }
-                    JWTPrincipal(credentials.payload)
-                } catch (e: Throwable) {
-                    null
-                }
-            }
-        }
         jwt("aad") {
             verifier(jwkProviderAad, configAad.metadata.issuer)
             validate { credentials ->
