+++ apps/etterlatte-proxy/src/main/kotlin/config/Config.kt
@@ -5,14 +5,10 @@ import io.ktor.client.call.body
 import io.ktor.client.request.get
 import io.ktor.server.config.ApplicationConfig
 import no.nav.etterlatte.routes.httpClientWithProxy
-import no.nav.etterlatte.routes.jsonClient
 
 data class Config(
     val sts: Sts,
     val aad: AAD,
-    val tokenX: TokenX,
-    val inntektskomponentenUrl: String,
-    val aaregUrl: String,
     val regoppslagUrl: String,
     val institusjonsoppholdUrl: String,
     val tilbakekrevingUrl: String
@@ -33,16 +29,6 @@ data class Config(
         }
     }
 
-    data class TokenX(
-        val metadata: Metadata,
-        val clientId: String,
-    ) {
-        data class Metadata(
-            @JsonProperty("issuer") val issuer: String,
-            @JsonProperty("jwks_uri") val jwksUri: String,
-        )
-    }
-
     data class AAD(
         val metadata: Metadata,
         val clientId: String,
@@ -55,9 +41,7 @@ data class Config(
 }
 
 suspend fun ApplicationConfig.load() = Config(
-    inntektskomponentenUrl = property("inntektskomponenten.url").getString(),
     institusjonsoppholdUrl = property("institusjonsopphold.url").getString(),
-    aaregUrl = property("aareg.url").getString(),
     regoppslagUrl = property("regoppslag.url").getString(),
     tilbakekrevingUrl = property("tilbakekreving.url").getString(),
     sts = Config.Sts(
@@ -71,9 +55,5 @@ suspend fun ApplicationConfig.load() = Config(
     aad = Config.AAD(
         metadata = httpClientWithProxy().use { it.get(property("aad.wellKnownUrl").getString()).body() },
         clientId = property("aad.clientId").getString()
-    ),
-    tokenX = Config.TokenX(
-        metadata = jsonClient().use { it.get(property("tokenx.wellKnownUrl").getString()).body() },
-        clientId = property("tokenx.clientId").getString()
     )
 )
